pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "student-search"
version="0.1.0"
requires-python = ">=3.13"
dependencies = [
    "gymnasium==1.2.2",
    "hydra-core==1.3.2",
    "numpy==2.3.5",
    "pygame==2.6.1",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "tensordict==0.10.0",
    "torch==2.9.1",
    "torchrl==0.10.1",
    "tqdm==4.67.1",
    "pettingzoo==1.24.3"
]

[project.optional-dependencies]
dev = [
    "black==25.12.0",
    "flake8==7.3.0",
    "pytest==9.0.2",
    "pre_commit==4.5.1"
]
